+++ src/utils/network.ts
@@ -1,0 +1,8 @@
+export function getNetwork(): string | undefined {
+  return import.meta.env.DFX_NETWORK;
+}
+
+export function isLocalNetwork(): boolean {
+  const network = getNetwork();
+  return !network || network === 'local';
+}
